Update Github_Tutorial.ino
because i want to learn work with git and githup i change it

+++ Github_Tutorial.ino
@@ -21,7 +21,7 @@ void setup()
 
 void loop() 
 {
-  int myValue = 0;
+  byte myValue = 0;
   myValue = analogRead(A0);
   
   Serial.print("The value is: ");
